[refactor] OperatorProvider 통합
