+++ src/main/java/com/jaeyeonling/calculator/type/Operator.java
@@ -2,12 +2,23 @@ package com.jaeyeonling.calculator.type;
 
 import com.jaeyeonling.calculator.BiOperation;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+
 public enum Operator {
     
     ADD("+", Integer::sum),
     SUBTRACT("-", (a, b) -> a - b),
     MULTIPLY("*", (a, b) -> a * b),
     DIVIDE("/", (a, b) -> a / b);
+
+    private static final Map<String, Operator> operatorMap = new HashMap<>();
+    static {
+        for (final Operator operator : Operator.values()) {
+            operatorMap.put(operator.symbol, operator);
+        }
+    }
 
     private final String symbol;
     private BiOperation<Integer> operation;
@@ -17,12 +28,17 @@ public enum Operator {
         this.operation = operation;
     }
 
-    public int apply(int a, int b) {
-        return operation.apply(a, b);
+    public static Operator symbol(final String symbol) {
+        final Operator operator = operatorMap.get(symbol);
+        if (Objects.isNull(operator)) {
+            throw new IllegalArgumentException();
+        }
+
+        return operator;
     }
 
-    public String getSymbol() {
-        return this.symbol;
+    public int apply(int a, int b) {
+        return operation.apply(a, b);
     }
 
 }
